requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.2.1
requests==2.32.5
slack_bolt==1.27.0
slack_sdk==3.39.0
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
gunicorn==21.2.0